refactor(Formula): 나눗셈 연산을 위한 Double형으로 변경

+++ src/main/kotlin/calculator/model/Formula.kt
@@ -9,10 +9,10 @@ class Formula(
         require(DIVIDE_BY_ZERO !in formulaString) { INVALID_DIVIDE_BY_ZERO_ERROR }
     }
 
-    fun converter(): Pair<List<Int>, List<String>> {
-        val numbers = number.findAll(formulaString).map { it.value.toInt() }.toList()
+    fun converter(): Pair<List<Double>, List<String>> {
+        val numbers = number.findAll(formulaString).map { it.value.toDouble() }.toList()
         val operators = formulaString.split(number).filter { it.trim().isNotEmpty() }
-        if (numbers.isEmpty() && operators.isEmpty()) return emptyList<Int>() to emptyList()
+        if (numbers.isEmpty() && operators.isEmpty()) return emptyList<Double>() to emptyList()
         require(numbers.size > operators.size) { INVALID_OPERATOR_INPUT_ERROR }
         return numbers to operators
     }
@@ -21,6 +21,7 @@ class Formula(
         private val number = Regex("\\d+")
         private val wrongString = Regex("[^\\d\\s+\\-*/]") // 숫자와 연산자가 아닌 글자
         private val wrongOperator = Regex("[+\\-*/]{2,}")
+
         private const val DIVIDE_BY_ZERO = "/0"
 
         private const val INVALID_STRING_INPUT_ERROR = "[ERROR] 문자가 입력됐습니다. 입력 형식을 확인해 주세요."
